New merge_update_part_of_file test
This adds a new test for a merge failure I have just found while working
on ATF.  This problem appears when running 'update' on a tree that has a
file that already carries part of the changes that are in the target
revision.  These already-committed changes confuse the merge algorithm
because the removed text is very similar to the text that is left.

This check fails with the current version of monotone and I have no fix
for it.  Hopefully having this test in the tree will force someone
knowledgeable enough to look at it, and it won't get lost.

Monotone-Parent: 59f939d5f1dd78058ba7b8752b110d76705d74da
Monotone-Revision: 941ca0c2752a096e4035c7366c9588c3a19f418f