Remove include of botan/bit_ops.h from gzip.cc
I have checked every relevant version of bit_ops.h, and it seems like
this header has never declared any function that is actually used by
gzip.cc (or at least its current incarnation).

This is a minor point, except that in 1.9.4 (upcoming botan dev
release), bit_ops.h is not available for applications anymore. It
still exists in the source tree, but is marked as an internal header
and so is not installed for general use.

Unrelated; add myself to AUTHORS on the basis of gzip.cc (modified
from botan's zlib wrapper) and the half-a-dozen-or-so small patches
I've written over time.

Monotone-Parent: 5596ea80ddef60062799f3945aeb8f408604b2da
Monotone-Revision: eab68914b6babe771529942206d269719d550772