* database.hh/cc: new method to create a new database if the one specified   by database_impl::filename does not exist * cmd_ws_commit.cc (setup), cmd_netsync.cc (clone): use that here; add a   FIXME that we still have to care about _MTN/options in a sane way, which   does currently not happen * lua_hooks.hh, lua_hooks.cc: add a new hook to retrieve the default   database alias; remove a couple of unneeded std:: prefixes * std_hooks.lua: implement the above hook and let it return ":default.mtn" * paths.hh: the path component for the internal database is no longer used
Whats missing right now is a sane way to resolve database aliases - the
method which does that currently should probably be hidden away in 
database_impl and database_impl itself should not only take a filename and
the db type, but an instance of the options and lua object to resolve the 
full path from the alias on its own. We should then also give database
a function to determine if a given external option set can be used to
resolve an aliased database to the same physical path (this is needed for
list databases).
On an unrelated note the (de)registration of known workspaces should be
moved from CMD(databases) into database as well and it should automatically
remove no longer existing or invalid workspaces from the list of known
workspaces. If we would then get some porcelain to register an existing
workspace as known workspace for a particular managed database, then this 
would be nice as well.

Monotone-Parent: d9e620538d12edbf56fbd13f6226643688ca274d
Monotone-Revision: 005d115dfa981a71a8314cf964d3ce00477ce097