New test for issue-120 - "ls changed PATH" should respect the given PATH filter
This test currently fails as the given PATH is not used to filter, and the
whole workspace is searched instead. The output of the test with 0.99.1
complains that the file is missing, even though it is in a different
directory to the one requested.

The fact that the whole workspace is being searched during "ls changed PATH"
explains why it is slower than "mtn status PATH", as "mtn status" respects
the PATH filter.

Monotone-Parent: c312be5d03b9ab6b4e2bebbd44b0b002743f173d
Monotone-Revision: cded246efa3cbe5357de11bcfb29c27d4b21026a