Print all certs as part of the revision header for "mtn log"
All the certs will be printed after the standard header of Author, Date,
Branch and before the Changelog and Comments.

Although "suspend" is a builtin cert acknowledged by monotone, currently
it will appear as part of the custom certs. Perhaps we could treat it like the
other builtin certs and provide formatting for it to as"Suspend:<spaces>"
like the others.

The custom certs only provide ": " as the separator and no extra spaces
because we don't know how long each cert name is, and it doesn't seem
worth calculating how much space would be needed to align them all.

NOTE: Should the ": " be in the translation files and ouput with _(": ") ?
"testresult" is a standard monotone cert but will appear along with the other
custom certs. Perhaps this should have special treatment like "suspend" as
well.
RE: the ": " separator between cert name and value

When the colorizer appears in the mainline the user can provide different
styles to help differentiate between cert name and value.

Monotone-Parent: 06738e14fae8309878e62cab9ea191be8fe891ce
Monotone-Revision: 694ed8ac92feff4643626d3c688f84425dba19a8