Make terminal_width for win32 consistent with unix and correct documentation
The unix version of terminal_width queries for the actual width, and on
failure returns 0; which means it's unlimited, see platform.hh. This is OK
as it's picked up on and corrected by guess_terminal_width which is used
throughout the code base and from Lua hooks.

The documentation for guess_terminal_width was slightly wrong as a width of
80 is only used on win32 systems; unix used 72, see default_terminal_width
in constants.hh.

NOTE:
Everywhere uses guess_terminal_width except for tick_write_count::write_ticks
which still uses terminal_width; I'm not sure at the moment if that's
intentional or an accident so I've left it.

Monotone-Parent: 46d59b9242bfcc6d147698168611cfa416ac62ea
Monotone-Revision: 2cb1b24a0643e108f8b2044449049053de601ac0